Auto-spoof mode for dnsmasq
Spoofing DNS responses or not is crucial to make the captive portal work:
- when in an online scenario where we route regular internet traffic,
we must not spoof or users will face the captive portal and not browse internet.
- when in offline scenario, returning valid (although spoofed) DNS responses
is mandatory for clients to make following http requests and reach the portal then
eventualy the hotspot.

those two modes are thus incompatible and the previous boolean configuration is fine
as long as the admin sets the correct mode and the environment doesn't change.

A realworld usecase though is a box that may or may not be connected to internet.
This `auto` mode allows this by toggling the DNS behavior based on the internet
connectivity status (provided by upstream/base-image)

This is done by calling a new conf-file from the main dnsmasq conf and pointing it
to /etc/dnsmasq-spoof.conf
Content of this file is different wether spoof was requested or not.
It's a single line of conf that enables spoofing when enabled and commented when not.

The IP returned when spoofing is an external IP (Cloudflare) on purpose as we want
offline clients to attempt real request so it traverses our router and get intercepted
and dnated to our captive portal

The toggling mechanism of `auto` attaches a systemd path watcher on /var/run/internet
and runs the toggle script based on its content.