Extracting checks from script logic
In order to reuse the configuration checks (for integration with other tools)
checks are deported to independant functions.

A new CheckResponse interface allows actual checking as well as returning a user-feedable message

In addition to extracting checks, a number of improvements were included:
- fixed setup.py using a static `1.0` version instead of version reported to scripts
- renamed module from `offspot_runtime_config` to `offspot_runtime`
- renamed `offspot_config_lib` sub-module to `configlib`
- most checks more resilient as checking input types
- timezones actual checked for existence instead of vague timezone-looking regex
- compose checking now includes (optional) image presence checking and exposure of ports
- ap.dhcp_range_for() now calculates an actual range instead of replacing strings
- ipv4 addresses are now checked for being actual IPv4 addresses with a flag to check if usable
- network checks now checks for valid network strings and compatibility with related IP address
- WiFi country code now checked for being an actual country code.
- Added support (and as default) for 00 country code meaning less-permissive radio options
- dhcp-range checks now checking for actual IP ranges with netwmask and host IP. ttl is also checked
- updated QA worflow
- unit tests for checks module (100% covered)
- test workflow

A couple new dependencies have been introduced to support better checks:
- tzdata that is used to check timezones
- iso3166 that is used to check country-codes (WiFi/AP)

What can (and probably will in a separate process) be improved is the wording of the feedback.
It's a bit difficult to settle on an harmonized, useful but not too verbose and repetitive template