Added checksum to Reader
Reader being a ~regular file, it deserves its own checksum and the previous handling
was not serving its purpose.

Without Checksum on Reader, we were only adding checksum to the File of the reader
and we had a special case to fetch it for download.kiwix.org but that could not work
(or was useless) since it required being on load-balancer and load-balancer is metalink.

It's now builder user's responsibility to create a Reader with a checksum or not.